+++ esp32c5/src/sha/m_mem.rs
@@ -12,17 +12,17 @@ impl W {}
 #[doc = "Sha M memory which contains message.\n\nYou can [`read`](crate::Reg::read) this register and get [`m_mem::R`](R). You can [`reset`](crate::Reg::reset), [`write`](crate::Reg::write), [`write_with_zero`](crate::Reg::write_with_zero) this register using [`m_mem::W`](W). You can also [`modify`](crate::Reg::modify) this register. See [API](https://docs.rs/svd2rust/#read--modify--write-api)."]
 pub struct M_MEM_SPEC;
 impl crate::RegisterSpec for M_MEM_SPEC {
-    type Ux = u8;
+    type Ux = u32;
 }
 #[doc = "`read()` method returns [`m_mem::R`](R) reader structure"]
 impl crate::Readable for M_MEM_SPEC {}
 #[doc = "`write(|w| ..)` method takes [`m_mem::W`](W) writer structure"]
 impl crate::Writable for M_MEM_SPEC {
     type Safety = crate::Unsafe;
-    const ZERO_TO_MODIFY_FIELDS_BITMAP: u8 = 0;
-    const ONE_TO_MODIFY_FIELDS_BITMAP: u8 = 0;
+    const ZERO_TO_MODIFY_FIELDS_BITMAP: u32 = 0;
+    const ONE_TO_MODIFY_FIELDS_BITMAP: u32 = 0;
 }
 #[doc = "`reset()` method sets M_MEM[%s] to value 0"]
 impl crate::Resettable for M_MEM_SPEC {
-    const RESET_VALUE: u8 = 0;
+    const RESET_VALUE: u32 = 0;
 }
